refactor: 실습 코드 패키지 이동
