+++ src/main/java/practice/Member.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package practice;
 
 public class Member {
