+++ src/main/java/practice/MemberDao.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package practice;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
